% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_angle_two_segments.R
\name{calc_angle_two_segments}
\alias{calc_angle_two_segments}
\title{Calculate the angle between two body segments.}
\usage{
calc_angle_two_segments(seg.df)
}
\arguments{
\item{seg.df}{a df of the coordinates of the two segments}
}
\value{
a new df with the angle between segments vs time
}
\description{
Calculate degrees between segments across frames.
   two segments are described by three keypoints
   "M" defines the end point of the segment to measure
   "P" defines the point to pivot about
       (it also defines one end of the measured segment
       and one end of the baseline segment)
   "B" defines the end point of the reference baseline
   seg.df contains the coordinates for these three keypoints
}
