% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_angular_kinematics.R
\name{calc_angular_kinematics}
\alias{calc_angular_kinematics}
\title{Calculate angular kinematics.}
\usage{
calc_angular_kinematics(settings, df)
}
\arguments{
\item{settings}{experimental settings}

\item{df}{the results df}
}
\value{
the original df with new columns added
}
\description{
Calculate angular kinematics using 'tail.curve.zero.rotate' in which
    tail curvature has been standardized for rotation (tail curve
    always moves in positive direction during stage 1) and adjusted
    for any baseline curvature offset.
}
