% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_linear_kinematics.R
\name{calc_linear_kinematics}
\alias{calc_linear_kinematics}
\title{Calculate linear kinematics.}
\usage{
calc_linear_kinematics(df, settings)
}
\arguments{
\item{df}{the results df}

\item{settings}{experimental settings}
}
\value{
the original df with new columns added
}
\description{
Using the keypoint specified in the settings file (usually S2),
    calculate the distance, displacement, speed, velocity, and
    acceleration of this keypoint across time.
}
