% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_morphology.R
\name{calc_morphology}
\alias{calc_morphology}
\title{Calculate body length and body bend..}
\usage{
calc_morphology(df, settings)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
the original df with body length and body bend added
}
\description{
Calculate the length of the larvae.
    Calculate body bend as the angle between
    the swim bladder segment and the first tail segment.
    Both calculations use the 10 pre-stimulus frames.
    Inserts these values into the original df.
}
