% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tail_curvature.R
\name{calc_tail_curvature}
\alias{calc_tail_curvature}
\title{Calculate tail curvature}
\usage{
calc_tail_curvature(df)
}
\arguments{
\item{df}{the results df}
}
\value{
the original df with tail.curve and body.curve
}
\description{
Calculate tail curvature as the sum of the angles
    of each tail segment. ALso calculate total body vurvature
    (tail curvatures and head curvature). Use smoothed curvature data
    in calculation.
}
