% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs_by_stage.R
\name{calcs_by_stage}
\alias{calcs_by_stage}
\title{Calculate kinematics on a stage-by-stage basis.}
\usage{
calcs_by_stage(df, settings, pk.vl.final.df)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}

\item{pk.vl.final.df}{a df containing the transitions between stages}
}
\value{
a new df consisting of a single row of kinematic values
}
\description{
Calculate overall distance, displacement, peak speed, and peak acceleration.
   Calculate appropriate kinematic variables for stage 1, 2, and 3.
   Compile all values into a new df (consisting of a single row for the
   particular trial under analysis.)
}
