% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_baseline_adj.R
\name{curvature_baseline_adj}
\alias{curvature_baseline_adj}
\title{curvature baseline adjustment}
\usage{
curvature_baseline_adj(df)
}
\arguments{
\item{df}{the results df (with a 'tail.curve.inv' column)}
}
\value{
the original df with a new 'tail.curve.inv.zero' column
}
\description{
Larvae often start with a baseline body curvature that is not zero,
    i.e. there is some positive or negative curvature offset. This
    offset must be removed prior to making calculations such as
    peak curvature, etc. This script adjusts curvature values to a zero starting baseline.
}
