% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_compile.R
\name{dd_compile}
\alias{dd_compile}
\title{Compile experiments}
\usage{
dd_compile(project, data_dir, results_dir, metadata_file)
}
\arguments{
\item{project}{a name for the project}

\item{data_dir}{the path to the data directory ("./" for current dir)}

\item{results_dir}{the path to the directory used for storing results}

\item{metadata_file}{the path to the metadata file}
}
\value{
a 'by ms' df and a 'mean' df
}
\description{
Assemble metadata and individual kinematic results into complete df's
}
