% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_kinematics.R
\name{dd_kinematics}
\alias{dd_kinematics}
\title{dd_kinematics}
\usage{
dd_kinematics(
  analysis_mode,
  single_video_ID,
  DLC_data_dir,
  settings_file,
  kin_results_dir,
  plots,
  pk_vl_save
)
}
\arguments{
\item{analysis_mode}{either "single", "batch" or "DAQ"}

\item{single_video_ID}{for single analysis mode, the ID of the trial}

\item{DLC_data_dir}{path to the DLC .csv pose estimate directory}

\item{settings_file}{path to the settings.txt file}

\item{kin_results_dir}{path to a directory for kinematic results}

\item{plots}{TRUE to make plots, FALSE to suppress plots}

\item{pk_vl_save}{TRUE to create pk_vl_df.csv, FALSE to suppress}
}
\value{
parameters to start kin_analysis
}
\description{
This is the starting script that assembles information
for the type of kinematic analysis, then calls kin_analysis
}
