% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_pks_vls.R
\name{detect_pks_vls}
\alias{detect_pks_vls}
\title{Detect peaks and valleys}
\usage{
detect_pks_vls(df, settings)
}
\arguments{
\item{df}{the results df}

\item{settings}{experimental settings}
}
\value{
a df identifying peaks and valleys
}
\description{
what function does (sentence case, no period)
}
