% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv_data.R
\name{get_csv_data}
\alias{get_csv_data}
\title{Open the .csv coordinate file}
\usage{
get_csv_data(video.filename, video.ID, data_dir, settings)
}
\arguments{
\item{data_dir}{The data directory path. This is most likely "./" when using
to evaluate current video.}

\item{settings}{The setting for this experiment.}
}
\value{
a results df that forms the starting point for all subsequent analyses
}
\description{
Open the .csv output coordinate data file from DeepLabCut.
    Obtain a video.ID from the .csv filename.
    Read the data and create an output dataframe.
}
