% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segment_coordinates.R
\name{get_segment_coordinates}
\alias{get_segment_coordinates}
\title{Calculate the angle between two body segments.}
\usage{
get_segment_coordinates(df, settings, segment)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}

\item{segment}{a segment (ex. "head" or "tail1")}
}
\value{
the seg.df used for input into 'calc_angle_two_segments'
}
\description{
Specify a segment, and this function will get the obtain
   the mm values from the results df and rename than for input
   into the 'calc_angle_two_segments' function.
}
