% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_settings}
\alias{get_settings}
\title{Get settings for the experiment.}
\usage{
get_settings(settings_path)
}
\arguments{
\item{settings_path}{the path to the settings txt file}
}
\value{
the settings data
}
\description{
Open the settings file, add variables, make calibration calculations,
    and return the setting df used when analyzing this particular experiment.
}
