% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin_process.R
\name{kin_process}
\alias{kin_process}
\title{Process current .csv file.}
\usage{
kin_process(
  video.filename,
  video.ID,
  settings_path,
  data_dir,
  results_dir,
  plots,
  pk_vl_save
)
}
\arguments{
\item{settings_path}{The path to your 'settings.txt' file.
This file contains important information about your experiment.
It is normally in a sub-directory labelled 'info'. Therefore, this parameter
is usually "./info/settings.txt".}

\item{data_dir}{The path to the current .csv file you want to analyze.
Normally, this is the .csv file that was just analyzed by DeepLabCut,
i.e. your current video. Therefore, this path is usually the current
working directory, or "./". Note that only ONE .csv file can be in this
directory.}

\item{results_dir}{The sub-directory where you want the df's created to be store.
An example would be "./results", a sub-directory of your working directory.}
}
\value{
Three df's, a panel of plots, and a summary table.
}
\description{
Process the DeepLabCut output .csv file that is in the current
    working directory.
      1. convert x, y pixel coordinates into mm coordinates
      2. calculate angular and linear kinematics
      3. identify the three stages of the escape response
      4. collapse multiple trials per fish into a mean value per fish
      5. make plots of selected responses
      6. make a summary table of selected variables
      7. save three df's: 'by_ms', 'one_row', and 'pk_vl'
    Your working directory should be setup as follows:
      1. A sub-directory named 'info'. Within this sub-directory a
         file labeled 'settings.txt' that contains information about
         the experiment.
      2. A data directory which contains a SINGLE .csv file,
         the file you want to analyze. During routine data collection
         this directory is your working directory. This parameter can
         be modified for cases where the experiment for analysis may
         be in a different directory.
      3. A directory where the df's created will be stored.
         This is usually a sub-directory named "results'.
}
