% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xkin_setup.R
\name{kin_setup}
\alias{kin_setup}
\title{Setup for processing.}
\usage{
kin_setup(
  analysis.mode,
  data_dir,
  settings_file,
  results_dir,
  video.ID.no,
  plots
)
}
\arguments{
\item{data_dir}{path to directory holding multiple videos (ignored for batch = FALSE)}

\item{settings_path}{the path to the video settings file}

\item{batch}{analyze either a single video (batch = FALSE) or multiple videos (batch = TRUE)}

\item{data_results}{path to directory where results are stored (ignored for batch = FALSE)}
}
\value{
selects video file(s) and initiates analysis
}
\description{
Set-up for processing either a single video or multiple videos.
}
