% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{make_plots}
\alias{make_plots}
\title{Make the final plot}
\usage{
make_plots(df, settings, row.df, video.ID)
}
\arguments{
\item{df}{the results df}

\item{settings}{the exp settings}

\item{row.df}{the row.df of results for the plot}

\item{video.ID}{the video ID for the plot}
}
\value{
a composite plot of kinematic results
}
\description{
Assemble individual plots into a final plot.
}
