% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary.R
\name{make_summary}
\alias{make_summary}
\title{Make a summary of results}
\usage{
make_summary(df, settings, row.df)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}

\item{row.df}{overall and stage-specific kinematic df}
}
\value{
summary.df, a df with selected kinematic variables for use in plots
}
\description{
Compiles a summary of results, including
    overall diatance and displacement,
    peak velocity and accelration, and
    tail beat frequency and amplitude.
}
