% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_fish.R
\name{number_fish}
\alias{number_fish}
\title{Number fish}
\usage{
number_fish(info.df)
}
\arguments{
\item{info.df}{the info.df created by kin_compile}
}
\value{
the info.df with the new column 'fish.no'
}
\description{
Consecutively number fish for a more compact way to identify individual fish
}
