% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_vl_calcs.R
\name{pk_vl_calcs}
\alias{pk_vl_calcs}
\title{Identify kinematic variables at each stage transition.}
\usage{
pk_vl_calcs(df, pk.vl.df.adj, settings)
}
\arguments{
\item{df}{the results df}

\item{pk.vl.df.adj}{a df with tail curve peaks and valleys}
}
\value{
pk.vl.final.df, a df with stage transition values
}
\description{
This function takes the 'pl_vl_df' and numbers the
    stage transition points and calculates variables
    at each stage transition. These values are then used
    in subsequent functions.
}
