% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acc.R
\name{plot_acc}
\alias{plot_acc}
\title{Plot acceleration vs movement time.}
\usage{
plot_acc(df, settings, ER_plot_theme)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
a plot
}
\description{
Create an acceleration plot using
    the results df.
}
