% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_displ.R
\name{plot_dist_displ}
\alias{plot_dist_displ}
\title{Plot distance and displacement vs movement time.}
\usage{
plot_dist_displ(df, settings, ER_plot_theme)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
a plot
}
\description{
Create a distance and displacement plot using
    the results df.
}
