% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likelihood.R
\name{plot_likelihood}
\alias{plot_likelihood}
\title{Plot likelihood vs movement time.}
\usage{
plot_likelihood(df, settings, video.ID)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
a plot
}
\description{
Create a multi-panel likelihood plot using
    the results df. Each panel
    displays a different segment.
}
