% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spd_vel.R
\name{plot_spd_vel}
\alias{plot_spd_vel}
\title{Plot speed and velocity vs movement time.}
\usage{
plot_spd_vel(df, settings, ER_plot_theme)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
a plot
}
\description{
Create a speed and velocity plot using
    the results df.
}
