% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tail_ang_vel.R
\name{plot_tail_ang_vel}
\alias{plot_tail_ang_vel}
\title{Plot tail angular velocity vs movement time.}
\usage{
plot_tail_ang_vel(df, settings, ER_plot_theme)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
a plot
}
\description{
Create a tail angular velocity plot using
    the results df.
}
