% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tail_curve.R
\name{plot_tail_curve}
\alias{plot_tail_curve}
\title{Plot tail curvature vs movement time.}
\usage{
plot_tail_curve(df, settings, ER_plot_theme, row.df)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
a plot
}
\description{
Create a tail curvature plot using
    the results df. Plot the 'tail.curve.inv.zero' data,
    not the raw "tail.curve' data.
}
