% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tail_segments.R
\name{plot_tail_segments}
\alias{plot_tail_segments}
\title{Plot tail segment angles vs movement time.}
\usage{
plot_tail_segments(df, settings, ER_plot_theme)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
a plot
}
\description{
Create a plot of tail angle for each tail segment using
    the results df.
}
