% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_adj_pk_vl.R
\name{rotation_adj_pk_vl}
\alias{rotation_adj_pk_vl}
\title{Adjust rotation direction in pk.vl.df.thres}
\usage{
rotation_adj_pk_vl(df, rot.adj)
}
\arguments{
\item{df}{the df}

\item{rot.adj}{logical variable for whether a rotational adjustment is needed}
}
\value{
output description
}
\description{
Adjust pk.vl.df.thres
}
