% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_adj_response.R
\name{rotation_adj_response}
\alias{rotation_adj_response}
\title{Adjust rotation direction in main.df}
\usage{
rotation_adj_response(df, rot.adj)
}
\arguments{
\item{df}{the df}

\item{rot.adj}{logical variable for whether a rotational adjustment is needed}
}
\value{
output description
}
\description{
Adjust rotation direction in main.df
}
