% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_adj_test.R
\name{rotation_adj_test}
\alias{rotation_adj_test}
\title{Rotational adjustment test}
\usage{
rotation_adj_test(df)
}
\arguments{
\item{df}{the df}

\item{rot.adj}{logical variable for whether a rotational adjustment is needed}
}
\value{
true (adjustment necessary) or false (no adjustment necessary)
}
\description{
Test whether a rotational adjustment in tail curvature is needed
}
