% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_speed.R
\name{smooth_speed}
\alias{smooth_speed}
\title{Smooth speed over time.}
\usage{
smooth_speed(df, settings)
}
\arguments{
\item{df}{the results df}

\item{settings}{the experimental settings}
}
\value{
the original df with a new column of smoothed angle data
}
\description{
Use loess to smooth the speed response.
}
