% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_pk_vl.R
\name{threshold_pk_vl}
\alias{threshold_pk_vl}
\title{Threshold small peaks and valleys}
\usage{
threshold_pk_vl(df, settings)
}
\arguments{
\item{df}{the df}

\item{settings}{the trial settings (needs threshold value)}
}
\value{
peak and valley df with small peaks and valleys ignored
}
\description{
Eliminate small peaks and valleys that may occur before the C-start based on threshold
}
